+++ src/main/java/com/smiley/smileybackend/dto/response/TestDateDto.java
@@ -23,7 +23,7 @@ public class TestDateDto {
     private int totalWearTimeExp;
 
     @ApiModelProperty( example = "획득 경험치(목표 착용 시간)")
-    private int objectiveWearTime;
+    private int targetWearTimeEXP;
 
     @ApiModelProperty( example = "획득 경험치(매거진 읽기)")
     private int magazineExp;
@@ -35,12 +35,12 @@ public class TestDateDto {
     private int commentExp;
 
     @Builder
-    public TestDateDto(int wearTime, int readMagazineNumber, int totalExp, int totalWearTimeExp, int objectiveWearTime, int magazineExp, int badgeExp, int commentExp) {
+    public TestDateDto(int wearTime, int readMagazineNumber, int totalExp, int totalWearTimeExp, int targetWearTimeEXP, int magazineExp, int badgeExp, int commentExp) {
         this.wearTime = wearTime;
         this.readMagazineNumber = readMagazineNumber;
         this.totalExp = totalExp;
         this.totalWearTimeExp = totalWearTimeExp;
-        this.objectiveWearTime = objectiveWearTime;
+        this.targetWearTimeEXP = targetWearTimeEXP;
         this.magazineExp = magazineExp;
         this.badgeExp = badgeExp;
         this.commentExp=commentExp;
@@ -50,11 +50,11 @@ public class TestDateDto {
         SecureRandom random = new SecureRandom ();
         this.wearTime = random.nextInt(760)+200;
         this.readMagazineNumber = random.nextInt(3);
-        this.totalWearTimeExp = random.nextInt(16000);
-        this.objectiveWearTime = random.nextInt(11000);
+        this.totalWearTimeExp = random.nextInt(16000)*5;
+        this.targetWearTimeEXP = random.nextInt(11000);
         this.magazineExp = random.nextInt(12)*100;
         this.badgeExp = random.nextInt(14)*1000;
         this.commentExp=random.nextInt(20)*500;
-        this.totalExp = this.totalWearTimeExp+this.objectiveWearTime+this.magazineExp+this.badgeExp;
+        this.totalExp = this.totalWearTimeExp+this.targetWearTimeEXP+this.magazineExp+this.badgeExp;
     }
 }
